+++ node/rpc/src/gas_api.rs
@@ -12,9 +12,9 @@ use blockstore::BlockStore;
 use chain::{BASE_FEE_MAX_CHANGE_DENOM, BLOCK_GAS_TARGET, MINIMUM_BASE_FEE};
 use fil_types::{verifier::ProofVerifier, BLOCK_GAS_LIMIT};
 use message::{unsigned_message::json::UnsignedMessageJson, UnsignedMessage};
-use message::{ChainMessage, Message, MessageSendSpec};
+use message::{ChainMessage, Message};
 use num_bigint::BigInt;
-use rpc_api::data_types::RPCState;
+use rpc_api::data_types::{MessageSendSpec, RPCState};
 
 const MIN_GAS_PREMIUM: f64 = 100000.0;
